requirements.txt
requests
beautifulsoup4
pandas
supabase
yt-dlp
python-dotenv
aiohttp
schedule
google-generativeai
fuzzywuzzy
python-levenshtein
websockets>=11,<16
sqlalchemy
psycopg2-binary
